refactor: Move execute command logic to a separate module
This commit refactors the execute command by moving its implementation from the main application file to a dedicated module, enhancing code organization and maintainability. The new structure allows for better separation of concerns and prepares the codebase for future enhancements. The command retains its functionality, supporting various input sources and output formats while improving clarity and documentation.
